make MiniBeagle project more manageable
