+++ iOS/Example/MiniBeagle/MiniBeagle.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "group:Playground.playground">
+   </FileRef>
+   <FileRef
+      location = "group:MiniBeagle.xcodeproj">
+   </FileRef>
+</Workspace>
